# WIP: Fix type instability in groupby()