@nalimilan 2017-02-22
Closing in favor of https://github.com/JuliaData/DataTables.jl/pull/17.